move fund autocomplete into own file
